UserModule::ensure_user method added
Added a method that makes it possible for methods to run code in a block that tries to get the global username first if needed and run the provided code. If no user could be found it throws an exception. Neat way to get similar logic for all functions. Loving Ruby already :p